+++ src/frontend/apps/desk/src/features/mail-domains/mailboxes/api/index.tsx
@@ -1,2 +1,3 @@
 export * from './useCreateMailbox';
 export * from './useMailboxes';
+export * from './useUpdateMailbox';