+++ src/frontend/apps/desk/src/features/mail-domains/mailboxes/components/index.ts
@@ -1,3 +1,4 @@
 export * from './ModalCreateMailbox';
+export * from './ModalUpdateMailbox';
 export * from './MailBoxesView';
 export * from './panel';