+++ src/frontend/apps/desk/src/features/mail-domains/mailboxes/types.ts
@@ -14,3 +14,13 @@ export type MailDomainMailboxStatus =
   | 'disabled'
   | 'pending'
   | 'failed';
+
+export interface ViewMailbox {
+  id: string;
+  email: string;
+  first_name: string;
+  last_name: string;
+  local_part: string;
+  secondary_email: string;
+  status: MailDomainMailboxStatus;
+}